requirements.txt
absl-py==2.4.0
aiofiles==23.2.1
annotated-types==0.7.0
anyio==3.7.1
astunparse==1.6.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dnspython==2.8.0
fastapi==0.104.1
flatbuffers==25.12.19
gast==0.7.0
google-auth==2.48.0
google-auth-oauthlib==1.2.4
google-pasta==0.2.0
grpcio==1.78.0
h11==0.16.0
h5py==3.15.1
httptools==0.7.1
idna==3.11
keras==2.15.0
libclang==18.1.1
Markdown==3.10.2
MarkupSafe==3.0.3
ml-dtypes==0.2.0
numpy==1.24.3
oauthlib==3.3.1
opencv-python==4.8.1.78
opt_einsum==3.4.0
packaging==26.0
Pillow==10.1.0
protobuf==4.25.8
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.5.0
pydantic_core==2.14.1
pymongo==4.6.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.3.0
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.24.0
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.5
wrapt==1.14.2
